apps: move pkg/apps/cmd to pkg/oc
