+++ hack/import-restrictions.json
@@ -453,7 +453,6 @@
       "github.com/openshift/origin/pkg/apps/apis/apps/test",
       "github.com/openshift/origin/pkg/apps/client/internalversion",
       "github.com/openshift/origin/pkg/apps/client/v1",
-      "github.com/openshift/origin/pkg/apps/cmd",
       "github.com/openshift/origin/pkg/apps/prune",
       "github.com/openshift/origin/pkg/apps/registry/deployconfig",
       "github.com/openshift/origin/pkg/apps/util",
