--- pkg/apps/cmd/doc.go
@@ -1,3 +1,0 @@
-// Package cmd contains various interface implementations for command-line tools
-// associated with deploymentconfigs.
-package cmd
