+++ pkg/oc/cli/deploymentconfigs/OWNERS
@@ -1,0 +1,13 @@
+reviewers:
+  - kargakis
+  - smarterclayton
+  - mfojtik
+  - soltysh
+  - abhgupta
+  - tnozicka
+approvers:
+  - kargakis
+  - smarterclayton
+  - mfojtik
+  - soltysh
+  - tnozicka
