+++ pkg/oc/cli/deploymentconfigs/delete_test.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"reflect"
