+++ pkg/oc/cli/deploymentconfigs/generate.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"fmt"
