+++ pkg/oc/cli/deploymentconfigs/history.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"bytes"
