+++ pkg/oc/cli/deploymentconfigs/rollback.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"bytes"
