+++ pkg/oc/cli/deploymentconfigs/scale_test.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"testing"
