+++ pkg/oc/cli/deploymentconfigs/status.go
@@ -1,4 +1,4 @@
-package cmd
+package deploymentconfigs
 
 import (
 	"errors"
